remoteproc: qcom: q6v5: Fix VMID length
The VMID length has been fixed to u64 to accommodate all VMID's
in commit [1]. Change the parameter to be compatible to the API.

[1] https://lore.kernel.org/all/20221219225850.2397345-19-quic_eberman@quicinc.com/

+++ drivers/remoteproc/qcom_q6v5_pas.c
@@ -662,9 +662,9 @@ static int setup_mpss_dsm_mem(struct platform_device *pdev)
 	struct device_node *node;
 	struct resource res;
 	phys_addr_t mem_phys;
-	int curr_perm;
+	u64 curr_perm;
 	u64 mem_size;
-	int ret;
+	u64 ret;
 
 	newvm[0].vmid = QCOM_SCM_VMID_MSS_MSA;
 	newvm[0].perm = QCOM_SCM_PERM_RW;
@@ -700,9 +700,9 @@ static int setup_global_sync_mem(struct platform_device *pdev)
 	struct device_node *node;
 	struct resource res;
 	phys_addr_t mem_phys;
-	int curr_perm;
+	u64 curr_perm;
 	u64 mem_size;
-	int ret;
+	u64 ret;
 
 	curr_perm = BIT(QCOM_SCM_VMID_HLOS);
 	newvm[0].vmid = QCOM_SCM_VMID_HLOS;
